requirements.txt
numpy==1.20.2
Pillow==8.2.0
pygame==2.0.1
scipy==1.6.2
termcolor==1.1.0
